+++ src/crosschain-data/utils/QuorumManager.sol
@@ -8,7 +8,7 @@ abstract contract QuorumManager is IQuorumManager {
     /*///////////////////////////////////////////////////////////////
                             STATE VARIABLES
     //////////////////////////////////////////////////////////////*/
-    mapping(uint16 srcChainId => uint256 quorum) internal requiredQuorum;
+    mapping(uint64 srcChainId => uint256 quorum) internal requiredQuorum;
 
     /*///////////////////////////////////////////////////////////////
                             EXTERNAL FUNCTIONS
@@ -16,14 +16,14 @@ abstract contract QuorumManager is IQuorumManager {
     /// @dev NOTE: ability to batch set quorum
 
     /// @inheritdoc IQuorumManager
-    function setRequiredMessagingQuorum(uint16 srcChainId_, uint256 quorum_) external virtual {}
+    function setRequiredMessagingQuorum(uint64 srcChainId_, uint256 quorum_) external virtual {}
 
     /*///////////////////////////////////////////////////////////////
                             PUBLIC FUNCTIONS
     //////////////////////////////////////////////////////////////*/
 
     /// @inheritdoc IQuorumManager
-    function getRequiredMessagingQuorum(uint16 srcChainId_) public view returns (uint256 quorum_) {
+    function getRequiredMessagingQuorum(uint64 srcChainId_) public view returns (uint256 quorum_) {
         /// @dev no chain can have chain id zero. (validates that here)
         if (srcChainId_ == 0) {
             revert Error.INVALID_INPUT_CHAIN_ID();
